+++ Src/CSharpier.Rider/src/main/java/com/intellij/csharpier/CSharpierSettings.java
@@ -49,6 +49,16 @@ public class CSharpierSettings
         this.useServer = useServer;
     }
 
+    private boolean skipCliExePath;
+
+    public boolean getSkipCliExePath() {
+        return this.skipCliExePath;
+    }
+
+    public void setSkipCliExePath(boolean skipCliExePath) {
+        this.skipCliExePath = skipCliExePath;
+    }
+
     @Override
     public CSharpierSettings getState() {
         return this;